packages/enterprise-bar-construct/pyproject.toml
[project]
name = "enterprise-bar-construct"
version = "2.1.1"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.244.0,<3.0.0",
    "enterprise-constructs-base",
]
